reworked loop and changed references to be consistent
elsewhere in this code, "fileList" is used for a list of files.
Updated "fileNames" to "fileList" to reflect that.  Also changed the
loop to be more efficent and less confusing.  While the previous
version removed all existing files from the list each time through the
surrounding loop, the rewritten code is functionally equivalent, since
all files end up having to be waited for.