Remove dependency on external Logger process
Previously, orchestration handled sending log
messages to a database, and send a message to itself
signalling the Logger had completed all its work since
the pipelines wouldn't be able to send any more messages.

Logging hasn't been handled this was some time, and
the Logger process has only been used as a signal that
the processing has ended. Since this is really overkill for what needed
to be done, the Logger process is being removed.

This change removes all references to the logger managers list.  It
also removes the mechanism that sent events to the Logger to indicated
processing had completed.

Shutdown is now handled internally, and can be triggered (as before)
with the shutProd.py command.