Removed remnants of pex_policy
removed:

killcondor.py - used the old style pex_policy job description to identify and kill jobs.
writeNodeList.py - this was required in an early version of ctrl_orca to enumerate the nodes used (in pbs jobs, I believe).
ProvenanceRecorder.py - This was used to make database records for policy files used in a run.  This was in addition to the provenance information that used to be recorded via the provenance package, which is also obsolete, since it also used pex_policy.  I don't think a replacement was ever made for that package.

lsst/ctrl/orca/PolicyUtils.py - used to get all file names listed in a policy

Deleted line from ups/ctrl_orca.table, since pex_policy, is no longer required.

Removed "-r" option from bin.src/orca.py which referred to policy directory. This was an obsolete option.